wip: run into rust compiler bug???
I cannot, for the life of me, see how this is an issue.

line 153-154 of functional_macros/src/nodes/coherence.rs. If either of
those are made to borrow from `node`, we run into a compiler error where
`bad_edges` now contains a borrow from `nodes`??? iirc the types of it
and `tys` don't change, but somehow the compiler is propagating extra
information about borrows ig. I can't seem to make a MWI that reproduces
this. Very odd.