only check against direct edges when filtering
If we check against _all_ edges_c when filtering out edges_b, then we
may be overly aggressive in filtering edges near the top of the lattice;
what we should do instead is, only iterate over edges_c that are direct
edges. Those are the only ones that _must_ be transformed when another
edge_b is being transformed.