TryFunctor codegen
trying to re-use as much as possible from the Functor codegen. Still
running into a few issues putting it into practice (inside the
`nodes!{}` macro seems to be a bit strange wrt struct detection?) but
otherwise Seems Good

There is something I'd like to do later, which is make it so that the
output is constructed in-place in a MaybeUninit, so that the errors can
do funny control flow things, but that requires a bigger rewrite of the
macro. Maybe it'll be worth it later? but for now I'm already a bit
behind on work done vs work needing to be blogged about lol.