make nodes! and functional_macros::ast play nice
Previously, nodes! was using a helper node! to generate each of its
cases. This was convenient because it meant I didn't have to put
everything in one big case, and could rely on macro_rules!'s built-in
case finder to emit the correct code. However, this meant that what
functional_macros::ast saw was just a blob of macro invokations, which
meant it didn't work.

To work around this issue, I had to move everything into one big case
that generates all the structs at once, so that we can apply
functional_macros::ast to just the final product. Honestly? It wasn't as
bad as I expected!