ch5: validate
As expected, TryFunctor is _really handy_ here. All the time I spent
making a framework for it is paying off!!

Of course, there was still an error with the code it generated: I was
`assume_init` on uninitialized memory, which in release mode caused
various badness. Fortunately, my overall approach was still correct,
just had to rejigger a few things.

Onward!