ch7 validate
A bit trickier! I _really_ didn't want to have to write all that
recursive descent code myself, so I made some changes to Functor and
TryFunctor. They now take in a RecursiveCall parameter to change their
behavior when they encounter a node that would otherwise be recursive.
This way, we can control the recursion a bit better, making things just
a bit nicer, though still not 100% as nice as I'd like, though maybe
that's for a future project... Automatic visitor generation??