wip: another functional_macros bug...
This time, it's the following example:

```rust
struct List<T> {
    val: T,
    next: Box<List<T>>,
}
```

Currently, we're not able to detect that generating a
`List<TInput> -> List<TOutput>` implementation will conflict with the
`val` field. That's bad! And another thing that I thought of a while ago
(for the blog post) but still haven't implemented: making it so, when
conflicts like these are detected, doing a standard `List<T> -> List<T>`
implementation instead. So I'll need a third `_tys` entry on `AField`...

Anyways! lots of good stuff here. I prob won't work on this for a while
tho lol.