wip: oops, need to open a compiler bug report
i added the thing to specialize on Span<T>, but that seems to have
caused an infinite loop somewhere in the compiler. I'm not on the
latest, so I'll try the latest stable and latest nightly to see if
they're any different.

That's a good a stopping point as any I think!