wip: fix unresolved variable error
the solution was staring me in the face the entire time: just make
idents a span, u silly. I didn't want to do this because it was
redundant data, but with no way to access it where i needed it, who's to
say it's actually redundant? anyways, it's there now.

i wanted to do fancy things with functor but Rust won't let me... so
mean... making sure traits have a well-defined resolution order... give
me specialization already...