wip: revert extern node progress
So, the reason things don't work, is because if you have a structure
like this:

    struct Exp {
        exp: Option<Box<Exp>>,
    }

And you try to generate a bunch of implementation like this:

    impl Functor<Option<Box<Exp>>> for Exp { ... }
    impl Functor<Box<Exp>> for Exp { ... }
    impl Functor<Exp> for Exp { ... }

Let's actually think about what that second implementation would look
like:

    impl Functor<Box<Exp>> for Exp {
        type Input = Box<Exp>;
        type Mapped = Exp;
        fn fmap(self, f: ...) -> Exp {
            Exp {
                exp: self.exp.fmap(f), // Resolution happens here
            }
        }
    }

This tries to instantiate `Option<Box<Exp>>: Functor<Box<Exp>>`,
which then tries to find `Box<Exp>: Functor<Box<Exp>>`, which it cannot
find.

In fact, without specialization, it cannot possibly find this! `Box<T>`
already has a generic `Functor` implementation, and choosing between the
"base" implementation and the "generic" implementation necessarily
requires specialization if we want to keep the "generic" implementation
around at all (which we do). So! Time to give up again :)

This reverts commit be87149463271a884a8362008736576071b256cc.

This reverts commit d7c1c0567fb174c240e77df5452efc56b75ff968.