wip: Visit trait passes basic test
Thinking about it more, I'm probably going to need a different approach
for VisitMut. Some sort of FoldableMut is necessary perhaps? Probably
not possible, however, because the accumulator type could do something
w/ the mut ref and Rust will complain about that.

Wondering if I'll need to break out Functor for it? Like do the map
in-place with no recurring by default. That _might_ work but we'll see
how the borrows work out.