wip: rewrite variable_resolution to use VisitMut
this is exactly what I had in mind when I set out on this project;
the logic there is a _lot_ more concise to express than with TryFunctor.
I also re-wrote loop_labeling while I was at it, because that also has
some stateful stuff that makes more sense to express with the visitor
pattern.

I may have a lot more support code necessary now (esp in the macros),
but the logic is clear & concise <3