ch8
I just had the tacky stage left, pretty simple from there yep.

switch statements are next!! This is an extra credit but I am doing all
of those. I think they are going to be a bit tricky, but absolutely
going to be worth it if I want to have a "real" C compiler and not just
a toy one.