wip: fix ch1 extra junk parse
The problem was: "multiplication" structs weren't resetting their `ts`
parameter on failure, which meant that `Vec::from_tokens` wasn't doing
that, which meant that the overall program parse wasn't doing that. This
suddenly came up because:

+ Previously, the top-level parse item was just a single function, and
  now it's a `Vec` of them.
+ There were no other test cases checking for extra tokens like this.

I've added a test cases internal to my compiler and also made sure it
passes that specific book test now.