Add .chain() method to generated Visit traits
Overall fairly simple: it just runs the two visitors back-to-back, and
this is fine because the tree should not be mutated through the shared
references. It is definitely _possible_ to break things this way, but
not likely I don't think.

I should also think if I want to do this for the VisitMut trait too...
The whole point of this chaining is so that we can run multiple analyses
in effectively one pass over the AST, but as it stands I am less certain
that multiple passes can coincide like this. Maybe they still can??

Also thinking about this more I realized that my current approach
doesn't even run the pass once like I'd hoped lmao, it just runs the
passes back-to-back from wherever they started. I will have to make it
so that there is a separate function for "doing the thing at the current
node" vs "recurring down from the current node", maybe even complete
with something that helps control recursion? idk that's not as flexible
as what we have currently, where different types of nodes can have
different recursions... I'll need to think about this more lol