make chaining work for VisitMut too
I did some thinking, and overall I think this approach will work. Just
need to evolve the trait definition slightly so there is a distinction
between "the part that does the work" and "the part that does the
recursion". This does mean the overall number of methods will triple,
but that's fine probably.