fix Visit{,Mut} bug
the recursive case needs to actually do the work, lol

While running this, I got an accurate count of how long I am actually
spending running the logic for in these traversals: before this PR
(where we were never actually doing logic in traverses, just traversing)
it took 51s to run the full test suite. After this PR, it takes 54s.

Next up will be testing how much of that is the traverse and how much is
the logic. I hope the traverse is what's actually taking long here...