wip: start using .chain() in semantic analysis
currently doesn't compile because I haven't changed the tests. Also
currently won't work at all because I haven't extracted the errors from
the visitors.

Overall I am not too happy with the ergonomics. It really needs to be as
easy as writing a TryFunctor-compatible method but we do require a fair
amount of boilerplate still unfortunately. I have some ideas on how to
improve this but nothing is there yet. That's for later.