wip: TryCoalesce
this is a crazy pile of hacks and I am surprised I was able to
accomplish it at all without somehow running up against specialization
again. (I did have to make that tuple type in order for it to work lol)

overall I'm fairly pleased with the end result. Wonder if it works
though...