fix missing idents from function arguments
This bug was caused by the type check visitor doing some "early
recursion" into the function's arguments before the ident mapper had a
chance to change them to the correct idents. We fix this by separating
out the recursion into its proper place, so the pass is ordered like
we'd expect.

Overall this visitor combination seems a little sus, because the
ordering really does matter a lot of the time, but also it's very cool
to have everything "just work" when it does.

I probably need more integration-style tests instead of the unit-style
tests I've been writing so far...