ch10 parse
...huh. I didn't really end up needing too many of the extended parse
features after all. the `#[derive(Spanned)]` was still quite useful, and
it _feels_ nicer to not have everything required to be in one big macro,
but (a) changes still take as long to recompute in rust-analyser
(because I'm still using a lot of other, smaller macros), and (b) I now
have to write many things twice :(

ah well live & learn, this was also fast enough to add a new thing to.